# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel-Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/admin/travels/{travel}/tours": {
            "post": {
                "tags": [
                    "Tour"
                ],
                "summary": "Add a new tour",
                "description": "Add a new tour",
                "operationId": "storeTour",
                "parameters": [
                    {
                        "name": "travelID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Create a new tour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TourRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TourResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/travels": {
            "post": {
                "tags": [
                    "Travel"
                ],
                "summary": "Add a new travel",
                "description": "Add a new travel",
                "operationId": "storeTravel",
                "requestBody": {
                    "description": "Create a new travel",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TravelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TravelResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name has already been taken."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name has already been taken."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/travels/{travel}": {
            "put": {
                "tags": [
                    "Travel"
                ],
                "summary": "Update travel",
                "description": "Update travel",
                "operationId": "updateTravel",
                "parameters": [
                    {
                        "name": "travelID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update this travel",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TravelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TravelResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name has already been taken."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name has already been taken."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign in",
                "description": "Login by email, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|0HpPOk1fnQoyQVZJLZmPXwzcYdqDwkaH8Tzb8vKa"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The provided credentials are incorrect",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The provided credentials are incorrect"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Tour": {
                "title": "Tour",
                "description": "Tour model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "UUid",
                        "type": "integer",
                        "format": "integer",
                        "example": "99eb9c87-79ab-491d-9bb7-a29e77e14e79"
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "format": "string",
                        "example": "Magnam totam illum."
                    },
                    "starting_date": {
                        "title": "starting date",
                        "description": "starting date",
                        "type": "string",
                        "format": "date",
                        "example": "2023-08-18"
                    },
                    "ending_date": {
                        "title": "ending date",
                        "description": "ending date",
                        "type": "string",
                        "format": "date",
                        "example": "2023-08-23"
                    },
                    "price": {
                        "title": "ending date",
                        "description": "ending date",
                        "type": "number",
                        "format": "float",
                        "example": "154.60"
                    }
                },
                "type": "object"
            },
            "Travel": {
                "title": "Travel",
                "description": "Travel model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "UUid",
                        "type": "integer",
                        "format": "integer",
                        "example": "99eb9c87-79ab-491d-9bb7-a29e77e14e79"
                    },
                    "slug": {
                        "title": "slug",
                        "description": "slug",
                        "type": "string",
                        "format": "string",
                        "example": "travel-to-japan"
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "format": "string",
                        "example": "Travel to Japan"
                    },
                    "description": {
                        "title": "description",
                        "description": "description",
                        "type": "string",
                        "format": "text",
                        "example": "Super good travel to Japan"
                    },
                    "number_of_days": {
                        "title": "number of days",
                        "description": "number of days",
                        "type": "integer",
                        "format": "integer",
                        "example": 20
                    },
                    "number_of_night": {
                        "title": "number of night",
                        "description": "number of night",
                        "type": "integer",
                        "format": "integer",
                        "example": 19
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "title": "Login request",
                "description": "Login request body data",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string",
                        "example": "user@box.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password",
                        "type": "string",
                        "example": "123123123"
                    }
                },
                "type": "object"
            },
            "TourRequest": {
                "title": "TourRequest",
                "description": "Tour request body data",
                "required": [
                    "name",
                    "price",
                    "startiing_date",
                    "ending_date"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name of the new tour",
                        "type": "string",
                        "example": "Amazing tour"
                    },
                    "price": {
                        "title": "price",
                        "description": "price of the new tour",
                        "type": "number",
                        "format": "float",
                        "example": "99.99"
                    },
                    "starting_date": {
                        "title": "starting_date",
                        "description": "start day of the new tour",
                        "type": "string",
                        "format": "date",
                        "example": "2023-08-16"
                    },
                    "ending_date": {
                        "title": "ending_date",
                        "description": "end day of the new tour",
                        "type": "string",
                        "format": "date",
                        "example": "2023-08-21"
                    }
                },
                "type": "object"
            },
            "TravelRequest": {
                "title": "TravelRequest",
                "description": "Travel request body data",
                "required": [
                    "name",
                    "desctiption",
                    "number_of_days"
                ],
                "properties": {
                    "is_public": {
                        "title": "is_public",
                        "description": "publish current travel",
                        "type": "boolean",
                        "example": "true"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of the new travel",
                        "type": "string",
                        "example": "Travel to Japan"
                    },
                    "description": {
                        "title": "description",
                        "description": "description of the new travel",
                        "type": "string",
                        "example": "Super good travel to Japan"
                    },
                    "number_of_days": {
                        "title": "number_of_days",
                        "description": "duration of the travel",
                        "type": "string",
                        "format": "date",
                        "example": "20"
                    }
                },
                "type": "object"
            },
            "TourResource": {
                "title": "TourResource",
                "description": "Tour resource",
                "properties": {
                    "data": {
                        "title": "data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tour"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TourResource"
                }
            },
            "TravelResource": {
                "title": "TravelResource",
                "description": "Travel resource",
                "properties": {
                    "data": {
                        "title": "data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Travel"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TravelResource"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}